requirements.txt
requests
python-dotenv
steam